refactor: optimize and reorder imports
