+++ src/components/MoviesCardList/MoviesCardList.jsx
@@ -1,5 +1,4 @@
 import './MoviesCardList.css';
-
 import { MoviesCard } from '../MoviesCard/MoviesCard';
 
 function MoviesCardList({ movies, savedMovies }) {
