+++ src/pages/NotFound/NotFound.jsx
@@ -1,4 +1,5 @@
 import { Link } from 'react-router-dom';
+
 import './NotFound.css';
 
 function NotFound() {
