+++ src/pages/Register/Register.jsx
@@ -1,5 +1,4 @@
 import './Register.css';
-
 import { AuthPage } from '../../components';
 
 function Register() {
